requirements.txt
psycopg2-binary
pandas
scikit-learn
openai~=1.2.3
flask
jsonify
requests